Version 1.0 ready!
ProjAnalysis is a tool that lets you analyze participation on
WikiProjects on the English Wikipedia. The tool will give you a list of
users who have edited pages in the project’s space (i.e. the
WikiProject page itself and its subpages) as well as the pages in the
project’s scope (as determined through talk page templates). The
resulting product is a CSV file that lists each user and the number of
edits made to the project’s space and to pages in scope.